+++ modules/config.py
@@ -1,0 +1,112 @@
+# coding=utf-8
+
+import os
+from pathlib import Path
+
+import yaml
+
+
+VERSION = "2.1.1"
+
+
+def load_config():
+    """加载配置文件"""
+    config_path = os.environ.get("CONFIG_PATH", "config/config.yaml")
+
+    if not Path(config_path).exists():
+        raise FileNotFoundError(f"配置文件 {config_path} 不存在")
+
+    with open(config_path, "r", encoding="utf-8") as f:
+        config_data = yaml.safe_load(f)
+
+    print(f"配置文件加载成功: {config_path}")
+
+    # 构建配置
+    config = {
+        "VERSION_CHECK_URL": config_data["app"]["version_check_url"],
+        "SHOW_VERSION_UPDATE": config_data["app"]["show_version_update"],
+        "REQUEST_INTERVAL": config_data["crawler"]["request_interval"],
+        "REPORT_MODE": config_data["report"]["mode"],
+        "RANK_THRESHOLD": config_data["report"]["rank_threshold"],
+        "USE_PROXY": config_data["crawler"]["use_proxy"],
+        "DEFAULT_PROXY": config_data["crawler"]["default_proxy"],
+        "ENABLE_CRAWLER": config_data["crawler"]["enable_crawler"],
+        "ENABLE_NOTIFICATION": config_data["notification"]["enable_notification"],
+        "MESSAGE_BATCH_SIZE": config_data["notification"]["message_batch_size"],
+        "BATCH_SEND_INTERVAL": config_data["notification"]["batch_send_interval"],
+        "FEISHU_MESSAGE_SEPARATOR": config_data["notification"][
+            "feishu_message_separator"
+        ],
+        "SILENT_PUSH": {
+            "ENABLED": config_data["notification"]
+            .get("silent_push", {})
+            .get("enabled", False),
+            "TIME_RANGE": {
+                "START": config_data["notification"]
+                .get("silent_push", {})
+                .get("time_range", {})
+                .get("start", "08:00"),
+                "END": config_data["notification"]
+                .get("silent_push", {})
+                .get("time_range", {})
+                .get("end", "22:00"),
+            },
+            "ONCE_PER_DAY": config_data["notification"]
+            .get("silent_push", {})
+            .get("once_per_day", True),
+            "RECORD_RETENTION_DAYS": config_data["notification"]
+            .get("silent_push", {})
+            .get("push_record_retention_days", 7),
+        },
+        "WEIGHT_CONFIG": {
+            "RANK_WEIGHT": config_data["weight"]["rank_weight"],
+            "FREQUENCY_WEIGHT": config_data["weight"]["frequency_weight"],
+            "HOTNESS_WEIGHT": config_data["weight"]["hotness_weight"],
+        },
+        "PLATFORMS": config_data["platforms"],
+    }
+
+    # Webhook配置（环境变量优先）
+    notification = config_data.get("notification", {})
+    webhooks = notification.get("webhooks", {})
+
+    config["FEISHU_WEBHOOK_URL"] = os.environ.get(
+        "FEISHU_WEBHOOK_URL", ""
+    ).strip() or webhooks.get("feishu_url", "")
+    config["DINGTALK_WEBHOOK_URL"] = os.environ.get(
+        "DINGTALK_WEBHOOK_URL", ""
+    ).strip() or webhooks.get("dingtalk_url", "")
+    config["WEWORK_WEBHOOK_URL"] = os.environ.get(
+        "WEWORK_WEBHOOK_URL", ""
+    ).strip() or webhooks.get("wework_url", "")
+    config["TELEGRAM_BOT_TOKEN"] = os.environ.get(
+        "TELEGRAM_BOT_TOKEN", ""
+    ).strip() or webhooks.get("telegram_bot_token", "")
+    config["TELEGRAM_CHAT_ID"] = os.environ.get(
+        "TELEGRAM_CHAT_ID", ""
+    ).strip() or webhooks.get("telegram_chat_id", "")
+
+    # 输出配置来源信息
+    webhook_sources = []
+    if config["FEISHU_WEBHOOK_URL"]:
+        source = "环境变量" if os.environ.get("FEISHU_WEBHOOK_URL") else "配置文件"
+        webhook_sources.append(f"飞书({source})")
+    if config["DINGTALK_WEBHOOK_URL"]:
+        source = "环境变量" if os.environ.get("DINGTALK_WEBHOOK_URL") else "配置文件"
+        webhook_sources.append(f"钉钉({source})")
+    if config["WEWORK_WEBHOOK_URL"]:
+        source = "环境变量" if os.environ.get("WEWORK_WEBHOOK_URL") else "配置文件"
+        webhook_sources.append(f"企业微信({source})")
+    if config["TELEGRAM_BOT_TOKEN"] and config["TELEGRAM_CHAT_ID"]:
+        token_source = (
+            "环境变量" if os.environ.get("TELEGRAM_BOT_TOKEN") else "配置文件"
+        )
+        chat_source = "环境变量" if os.environ.get("TELEGRAM_CHAT_ID") else "配置文件"
+        webhook_sources.append(f"Telegram({token_source}/{chat_source})")
+
+    if webhook_sources:
+        print(f"Webhook 配置来源: {', '.join(webhook_sources)}")
+    else:
+        print("未配置任何 Webhook")
+
+    return config
