+++ modules/utils.py
@@ -1,0 +1,116 @@
+# coding=utf-8  
+  
+import re  
+from datetime import datetime  
+from pathlib import Path  
+from typing import Dict, List, Tuple, Optional, Union  
+  
+import pytz  
+  
+  
+def get_beijing_time():
+    """获取北京时间"""
+    return datetime.now(pytz.timezone("Asia/Shanghai"))
+
+
+def format_date_folder():
+    """格式化日期文件夹"""
+    return get_beijing_time().strftime("%Y年%m月%d日")
+
+
+def format_time_filename():
+    """格式化时间文件名"""
+    return get_beijing_time().strftime("%H时%M分")
+
+
+def clean_title(title: str) -> str:
+    """清理标题中的特殊字符"""
+    if not isinstance(title, str):
+        title = str(title)
+    cleaned_title = title.replace("\n", " ").replace("\r", " ")
+    cleaned_title = re.sub(r"\s+", " ", cleaned_title)
+    cleaned_title = cleaned_title.strip()
+    return cleaned_title
+
+
+def ensure_directory_exists(directory: str):
+    """确保目录存在"""
+    Path(directory).mkdir(parents=True, exist_ok=True)
+
+
+def get_output_path(subfolder: str, filename: str) -> str:
+    """获取输出路径"""
+    date_folder = format_date_folder()
+    output_dir = Path("output") / date_folder / subfolder
+    ensure_directory_exists(str(output_dir))
+    return str(output_dir / filename)
+
+
+def check_version_update(
+    current_version: str, version_url: str, proxy_url: Optional[str] = None
+) -> Tuple[bool, Optional[str]]:
+    """检查版本更新"""
+    try:
+        proxies = None
+        if proxy_url:
+            proxies = {"http": proxy_url, "https": proxy_url}
+
+        headers = {
+            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36",
+            "Accept": "text/plain, */*",
+            "Cache-Control": "no-cache",
+        }
+
+        response = requests.get(
+            version_url, proxies=proxies, headers=headers, timeout=10
+        )
+        response.raise_for_status()
+
+        remote_version = response.text.strip()
+        print(f"当前版本: {current_version}, 远程版本: {remote_version}")
+
+        # 比较版本
+        def parse_version(version_str):
+            try:
+                parts = version_str.strip().split(".")
+                if len(parts) != 3:
+                    raise ValueError("版本号格式不正确")
+                return int(parts[0]), int(parts[1]), int(parts[2])
+            except:
+                return 0, 0, 0
+
+        current_tuple = parse_version(current_version)
+        remote_tuple = parse_version(remote_version)
+
+        need_update = current_tuple < remote_tuple
+        return need_update, remote_version if need_update else None
+
+    except Exception as e:
+        print(f"版本检查失败: {e}")
+        return False, None
+
+
+def is_first_crawl_today() -> bool:
+    """检测是否是当天第一次爬取"""
+    date_folder = format_date_folder()
+    txt_dir = Path("output") / date_folder / "txt"
+
+    if not txt_dir.exists():
+        return True
+
+    files = sorted([f for f in txt_dir.iterdir() if f.suffix == ".txt"])
+    return len(files) <= 1
+
+
+def html_escape(text: str) -> str:
+    """HTML转义"""
+    if not isinstance(text, str):
+        text = str(text)
+
+    return (
+        text.replace("&", "&amp;")
+        .replace("<", "&lt;")
+        .replace(">", "&gt;")
+        .replace('"', "&quot;")
+        .replace("'", "&#x27;")
+    )
